% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dengue_cases.R
\name{dengue_cases}
\alias{dengue_cases}
\title{Dengue activity estimates for the world}
\format{
A data frame with 6.263 rows and 3 columns:
\describe{
\item{year}{Year for observation}
\item{country}{Country for observation}
\item{dengue_count}{dengue cases of country per year}
}
}
\source{
from Google dengue Trends (http://www.google.org/denguetrends). Changed original dataset to only show yearly estimates and reformatted.
}
\usage{
dengue_cases
}
\description{
This dataset is a reformatted version of a dataset of weekly estimated cases for dengue in different countries
around the world in 2002-2015 (from Google datasets). This dataset contains the years, dengue cases and country.
}
