% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/disease_count.R
\name{disease_count}
\alias{disease_count}
\title{Counts cases of flu or dengue in a country.}
\usage{
disease_count(df, country_name, disease)
}
\arguments{
\item{df}{Dataframe containing a column with countries, and flu_count or dengue_count for different dates.}

\item{country_name}{Name of country in country column}

\item{disease}{"flu" or "dengue"}
}
\value{
The total sum of cases of chosen disease throughout the years for a specific country
}
\description{
Counts cases of flu or dengue in a country.
}
\examples{
\dontrun{
If you have a dataframe with a column country that contains different countries,
and a column with dengue/flu_count for different dates.
And u want to know the sum of flu cases throughout the years
for Bolivia you would use this:

disease_count(df = dataframe, country = "Bolivia", disease = "flu")
}
}
