% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_date.R
\name{get_date}
\alias{get_date}
\title{Get day,month or year in a column of dataframe with dates if in yyyy/mm/dd format.}
\usage{
get_date(dfcolumn, dmy)
}
\arguments{
\item{dfcolumn}{Column in dataframe which contains the dates}

\item{dmy}{Choose whether you want the days, months or years by filling this parameter with "day", "month" or "year".}
}
\value{
The days, months or years depending on what you chose.
}
\description{
Get day,month or year in a column of dataframe with dates if in yyyy/mm/dd format.
}
\examples{
x <- data.frame(date = '2023-08-24')
get_date(x$date, "year")

}
