% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_image.R
\name{load_image}
\alias{load_image}
\title{Load an image}
\usage{
load_image(path)
}
\arguments{
\item{path}{The path to your image.}
}
\value{
A loaded image
}
\description{
Load an image
}
\examples{
\dontrun{
Path to your image might be: figures/picture.png.
This is what you would use: load_image("figures/picture.png")
}
}
