% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_cases.R
\name{plot_cases}
\alias{plot_cases}
\title{Generates a plot of dengue or flu cases throughout the years in different countries.}
\usage{
plot_cases(df, disease, plot_title)
}
\arguments{
\item{df}{Dataframe with a column "year", "country" and "flu_count"/"dengue_count"}

\item{disease}{Choose whether you want a plot for "flu" or "dengue"}

\item{plot_title}{Title of plot}
}
\value{
A plot for dengue or flu cases throughout the years in countries
}
\description{
Generates a plot of dengue or flu cases throughout the years in different countries.
}
\examples{
\dontrun{
If you have a dataframe with a column country that contains a country
(or more countries), and a column with dengue/flu_count for different years.
And if you would want a plot for flu cases in Germany from 2015-2020,
you would use:

plot_cases(df = dataframe, disease = "flu", plot_title = "Flu cases in Germany from 2015-2020")
}
}
